--- PresentationLayer/MainPage2.xaml
@@ -1,57 +1,0 @@
-﻿<Page x:Class="EP_HSRlearnIT.PresentationLayer.MainPage2"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      mc:Ignorable="d" 
-      d:DesignHeight="300" d:DesignWidth="300"
-      Title="MainPage2">
-
-    <UniformGrid Background="White">
-
-        <Border Style="{StaticResource TileBorderStyle}">
-            <Grid PreviewMouseLeftButtonDown="OverviewScreen_Click" MouseEnter="MenuTile_OnMouseEnter" MouseLeave="MenuTile_OnMouseLeave" Background="{StaticResource BackAreaBrush}">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="*"/>
-                    <RowDefinition Height="2*"/>
-                    <RowDefinition Height="2*"/>
-                </Grid.RowDefinitions>
-                <Image Grid.Row="1" Source="Images/eye-icon.png" Stretch="Uniform"/>
-                <TextBlock Grid.Row="2" Text="Übersicht AES GCM" Style="{StaticResource TileContentStyle}"/>
-            </Grid>
-        </Border>
-        <Border Style="{StaticResource TileBorderStyle}">
-            <Grid PreviewMouseLeftButtonDown="StepByStep_Click" MouseEnter="MenuTile_OnMouseEnter" MouseLeave="MenuTile_OnMouseLeave" Background="{StaticResource BackAreaBrush}">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="*"/>
-                    <RowDefinition Height="2*"/>
-                    <RowDefinition Height="2*"/>
-                </Grid.RowDefinitions>
-                <Image Grid.Row="1" Source="Images/step-icon.png" Stretch="Uniform"/>
-                <TextBlock  Grid.Row="2" Text="Schritt für Schritt Anleitung" Style="{StaticResource TileContentStyle}"/>
-            </Grid>
-        </Border>
-        <Border Style="{StaticResource TileBorderStyle}">
-            <Grid PreviewMouseLeftButtonDown="EncryptionDecryption_Click" MouseEnter="MenuTile_OnMouseEnter" MouseLeave="MenuTile_OnMouseLeave" Background="{StaticResource BackAreaBrush}">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="*"/>
-                    <RowDefinition Height="2*"/>
-                    <RowDefinition Height="2*"/>
-                </Grid.RowDefinitions>
-                <Image Grid.Row="1" Source="Images/key-icon.png" Stretch="Uniform"/>
-                <TextBlock Grid.Row="2" Text="Ver- und Entschlüsselung" Style="{StaticResource TileContentStyle}"/>
-            </Grid>
-        </Border>
-        <Border Style="{StaticResource TileBorderStyle}">
-            <Grid PreviewMouseLeftButtonDown="DragAndDrop_Click" MouseEnter="MenuTile_OnMouseEnter" MouseLeave="MenuTile_OnMouseLeave" Background="{StaticResource BackAreaBrush}">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="*"/>
-                    <RowDefinition Height="2*"/>
-                    <RowDefinition Height="2*"/>
-                </Grid.RowDefinitions>
-                <Image Grid.Row="1" Source="Images/drag-icon.png" Stretch="Uniform"/>
-                <TextBlock Grid.Row="2" Text="Drag- und Drop - Spiel" Style="{StaticResource TileContentStyle}"/>
-            </Grid>
-        </Border>
-    </UniformGrid>
-</Page>
