+++ PresentationLayer/MainWindow.xaml.cs
@@ -23,7 +23,7 @@ namespace EP_HSRlearnIT.PresentationLayer
         public MainWindow()
         {
             InitializeComponent();
-            MainFrame.Navigate(new MainPage2());
+            MainFrame.Navigate(new MainPage());
             Application.Current.MainWindow = this;
         }
         #endregion
@@ -46,7 +46,7 @@ namespace EP_HSRlearnIT.PresentationLayer
             switch (item)
             {
                 case "Startseite":
-                    MainFrame.Navigate(new MainPage2());
+                    MainFrame.Navigate(new MainPage());
                     break;
                 case "AES-GCM Übersicht":
                     MainFrame.Navigate(new AesGcmOverviewPage());
